I added a plugin to the POM to package the app into a jar.
Also: I got rid of the ./ in the css url references to the images. It was
telling the css to look in the very root, and that only worked here.
Once the jar was out in the wild, it was looking in the wrong place for
the images. I'm pretty sure after one round of testing, that it will
find them straight away now. It was so odd because I thought I would
have to look up a few folders using `../` or maybe put the images
in a folder inside of the resources folder that was named the same thing
as the package was, but no! I have no idea why that didn't work, but it
didn't?what worked was just leaving them in the root of the resources
folder and then just asking for them like, `url('image.jpg')`. Crazy.